refactor: ConnectionContext
