+++ packages/plugins/EVM/src/state/Connection/connection.ts
@@ -39,7 +39,8 @@ import {
 } from '@masknet/web3-shared-base'
 import { Web3 } from '@masknet/web3-providers'
 import type { BaseContract } from '@masknet/web3-contracts/types/types.js'
-import { createContext, dispatch } from './composer.js'
+import { dispatch } from './composer.js'
+import { createContext } from './context.js'
 import { Providers } from './provider.js'
 import type { EVM_Connection, EVM_Web3ConnectionOptions } from './types.js'
 import { Web3StateSettings } from '../../settings/index.js'
