+++ packages/plugins/EVM/src/state/Connection/interceptors/MetaMask.ts
@@ -1,8 +1,7 @@
-import { EthereumMethodType, PayloadEditor } from '@masknet/web3-shared-evm'
-import type { Context, Middleware } from '../types.js'
+import { EthereumMethodType, PayloadEditor, ConnectionContext, Middleware } from '@masknet/web3-shared-evm'
 
-export class MetaMask implements Middleware<Context> {
-    async fn(context: Context, next: () => Promise<void>) {
+export class MetaMask implements Middleware<ConnectionContext> {
+    async fn(context: ConnectionContext, next: () => Promise<void>) {
         // Evoke the unlock popup when metamask-like is locked before send transaction or sign message.
         if (PayloadEditor.fromPayload(context.request).risky) {
             await context.connection.connect(context.requestOptions)
