+++ packages/plugins/EVM/src/state/Connection/interceptors/None.ts
@@ -1,7 +1,7 @@
-import type { Context, Middleware } from '../types.js'
+import type { ConnectionContext, Middleware } from '@masknet/web3-shared-evm'
 
-export class NoneWallet implements Middleware<Context> {
-    async fn(context: Context, next: () => Promise<void>) {
+export class NoneWallet implements Middleware<ConnectionContext> {
+    async fn(context: ConnectionContext, next: () => Promise<void>) {
         if (context.risky) {
             context.abort(new Error('No allowed.'))
         }
