+++ packages/plugins/EVM/src/state/Connection/interceptors/Popups.ts
@@ -1,11 +1,10 @@
 import { isUndefined, omitBy } from 'lodash-es'
-import { ErrorEditor } from '@masknet/web3-shared-evm'
+import { ErrorEditor, Middleware, ConnectionContext } from '@masknet/web3-shared-evm'
 import { ExtensionSite, getSiteType, isEnhanceableSiteType } from '@masknet/shared-base'
-import type { Context, Middleware } from '../types.js'
 import { SharedContextSettings } from '../../../settings/index.js'
 
-export class Popups implements Middleware<Context> {
-    async fn(context: Context, next: () => Promise<void>) {
+export class Popups implements Middleware<ConnectionContext> {
+    async fn(context: ConnectionContext, next: () => Promise<void>) {
         // Draw the Popups up and wait for user confirmation before publishing risky requests on the network
         if (context.risky) {
             const response = await SharedContextSettings.value.send(
