+++ packages/plugins/EVM/src/state/Connection/middleware/Translator.ts
@@ -1,19 +1,18 @@
-import { ChainId } from '@masknet/web3-shared-evm'
+import { ChainId, Middleware, Translator as ChainTranslator, ConnectionContext } from '@masknet/web3-shared-evm'
 import { Base } from '../translators/Base.js'
 import { Polygon } from '../translators/Polygon.js'
-import type { Context, Middleware, Translator as ChainTranslator } from '../types.js'
 
 /**
  * JSON RPC transactor for EVM chains.
  */
-export class Translator implements Middleware<Context> {
+export class Translator implements Middleware<ConnectionContext> {
     private base = new Base()
 
-    private translators: Partial<Record<ChainId, ChainTranslator>> = {
+    private translators: Partial<Record<ChainId, ChainTranslator<ConnectionContext>>> = {
         [ChainId.Mumbai]: new Polygon(),
     }
 
-    async fn(context: Context, next: () => Promise<void>) {
+    async fn(context: ConnectionContext, next: () => Promise<void>) {
         const translator = this.translators[context.chainId] ?? this.base
         if (translator.encode) await translator.encode(context)
         await next()
