+++ packages/plugins/EVM/src/state/Connection/translators/Base.ts
@@ -1,12 +1,19 @@
 import { BigNumber } from 'bignumber.js'
 import { toHex } from 'web3-utils'
 import { GasOptionType, isLessThan, toFixed } from '@masknet/web3-shared-base'
-import { addGasMargin, ChainId, chainResolver, formatWeiToGwei, PayloadEditor } from '@masknet/web3-shared-evm'
-import type { Context, Translator } from '../types.js'
+import {
+    addGasMargin,
+    ChainId,
+    chainResolver,
+    formatWeiToGwei,
+    PayloadEditor,
+    Translator,
+    ConnectionContext,
+} from '@masknet/web3-shared-evm'
 import { Web3StateSettings } from '../../../settings/index.js'
 
-export class Base implements Translator {
-    async encode(context: Context) {
+export class Base implements Translator<ConnectionContext> {
+    async encode(context: ConnectionContext) {
         const config = context.config
         if (!config || PayloadEditor.fromPayload(context.request).readonly) return
 
@@ -70,5 +77,5 @@ export class Base implements Translator {
         // #endregion
     }
 
-    async decode(context: Context) {}
+    async decode(context: ConnectionContext) {}
 }
