+++ packages/plugins/EVM/src/state/Connection/translators/Optimistic.ts
@@ -1,11 +1,11 @@
-import type { Context } from '../types.js'
+import type { ConnectionContext } from '@masknet/web3-shared-evm'
 import { Base } from './Base.js'
 
 export class Optimistic extends Base {
-    override async encode(context: Context) {
+    override async encode(context: ConnectionContext) {
         throw new Error('Method not implemented.')
     }
-    override async decode(context: Context) {
+    override async decode(context: ConnectionContext) {
         throw new Error('Method not implemented.')
     }
 }
