+++ packages/plugins/EVM/src/state/Connection/types.ts
@@ -33,61 +33,3 @@ export interface ERC1155Metadata {
     description: string
     image: string
 }
-
-export interface Context {
-    readonly account: string
-    readonly chainId: ChainId
-    readonly requestId: number
-    readonly writeable: boolean
-    readonly risky: boolean
-    readonly providerType: ProviderType
-    readonly method: EthereumMethodType
-    readonly connection: EVM_Connection
-    readonly owner?: string
-    readonly identifier?: ECKeyIdentifier
-    readonly requestOptions?: EVM_Web3ConnectionOptions
-    readonly message?: string
-    readonly userOperation?: UserOperation
-    readonly proof?: Proof
-
-    /**
-     * JSON RPC request payload
-     */
-    readonly request: JsonRpcPayload
-
-    /**
-     * JSON RPC response object
-     */
-    readonly response?: JsonRpcResponse
-
-    config?: Transaction
-    requestArguments: RequestArguments
-    result: unknown
-    error: RecognizableError | null
-
-    /**
-     * Resolve a request and write down the result into the context. Alias of end(null, result)
-     */
-    write: (result: unknown) => void
-
-    /**
-     * Reject a request and throw an error. Alias of end(error)
-     */
-    abort: (error: unknown, fallback?: string) => void
-
-    /**
-     * Seal a request by resolving or rejecting it.
-     */
-    end: (error?: Error | null, result?: unknown) => void
-
-    toJSON: () => object
-}
-
-export interface Middleware<T> {
-    fn: (context: T, next: () => Promise<void>) => Promise<void>
-}
-
-export interface Translator {
-    encode?(context: Context): Promise<void>
-    decode?(context: Context): Promise<void>
-}
